Adds a deterministic, open-sourced input pipeline.
In order to support determinism, this ensures that the file shuffle seed is set
in the read config for TFDS datasets [1], that a seed is supplied for example
shuffling in tf.data.Dataset.shuffle [2], and that preprocessing functions are also
supplied a seed. A subsequent change will enforce that all preprocessing
functions actually make use of that seed. The `deterministic_data` module in
the `clu` project [3] is used to help set this up.

References:
[1]: https://www.tensorflow.org/datasets/determinism
[2]: https://www.tensorflow.org/api_docs/python/tf/data/Dataset#shuffle
[3] https://github.com/google/CommonLoopUtils/blob/main/clu/deterministic_data.py
PiperOrigin-RevId: 396024189