Fixes a bug that was causing ViT training failures due to the fewshot code.
Models using the default `common_fewshot` config were failing with the error:

```
Cannot add tensor to the batch: number of elements does not match. Shapes
are: [tensor]: [374,500,1], [batch]: [377,500,1] [Op:IteratorGetNext]
```

within the data loading part of the fewshot code.

The reason is that for some of the fewshot datasets, additional keys aside from
`images` and `label` were still present after fewshot preprocessing, including
a segmentation mask. Since the images are different sizes prior to
preprocessing, and the preprocessing only applies to the images, the
segmentation masks still had different sizes after preprocessing, and
therefore could not be batched. We shouldn’t be keeping those anyway, so the
fix is to include a `keep("image", "label")` in the fewshot config.

PiperOrigin-RevId: 421420507