Change the TFDS mocking number of examples in prep for the val iterator CL.
Currently, `tfds.testing.mock_data` ignores split slices, such as `"train[:9]",
and will instead always return a dataset with a number of examples equal to the
value of its `num_examples` parameter. A subsequent CL is changing the way the
validation dataset iterators are created and consumed, and this is resulting in
false test failures.

PiperOrigin-RevId: 421828032