Recreate the val iterator at eval time to improve performance after restarts.
The training scripts are tested to be reproducible even if they are stopped and
started again mid-way through training. In these restart cases, the scripts
will load the previous checkpoint, advance all iterators (including the dataset
iterators) to the correct step, and then continue training from that point.
This means the scripts will yield the same results whether they are run to
completion in one run, or stopped and started again along the way.

One potential drawback is the computational cost of advancing the dataset
iterators to the current step after these restarts, since all dataset loading
and preprocessing computation must be run for each batch. We have recently run
into issues where this cost is now a blocker to training in the presence of
preemptions.

Previously, both the training and validation datasets were set up as
infinitely-repeating datasets that were advanced by use of step counts. It is
unnecessary for the validation dataset to be implemented in this way, so this
CL instead creates a validation dataset with one epoch and then recreates an
iterator during each validation evaluation. The benefit is that there is no
longer a validation iterator that needs to be advanced after script restarts
described above.

A subsequent CL will address the training dataset iterator.

PiperOrigin-RevId: 421841768