Update Het+BE+GP's BE component.
Here, BE is applied at layers 21-23 instead of just 21+23, has bias terms, BE for representation_size layer, and BE at head layer.

For simplicity, I removed the ability to turn off BE and turn off Het. It added too much if-else branching, particularly with BE's extra head layer. Given we already have scripts for Het-only, BE-only, GP-only, and Het+GP, this script can just serve as Het+BE and Het+BE+GP.

Bugs:
+ Fixed an issue with representation_fn. It must use pre_ens_logits, not pre_logits.
PiperOrigin-RevId: 427926257