Fixes the TPU HBM OOM issue for BatchEnsemble with the active learning script.
Previously, we could only run the active learning script with the BE model
using batch_size=128, while the deterministic model could run with
batch_size=256. There were two issues: donate_argnums on the update_fns used
for the AL script were incorrect, and (more importantly) the AL script was
inadvertently disabling the newer JAX TPU runtime internally, which disabled
automatic defragmentation on the TPUs. That was leading to the HBM OOM issue
specifically for the AL script vs. the other scripts. The fix allows us to
increase the batch_size to 256 for BE in order to match the deterministic
model.
PiperOrigin-RevId: 437039351