Clean up, generalize, and test the code that computes reliability scores based on RDL experimental results.
This CL introduces some major changes in how ranks and scores are computed; the biggest difference between colab behavior and behavior in the checked-in code stems from how missing values are treated.

The previous code
1. Selects methods that report all metrics (`df.dropna(how='any')`)
2. For each category (prediction / uncertainty / adaptation), computes the ranking for methods that report those metrics (`df[categories].dropna(how='any')`)
3. Updates the rankings for methods in step 1. based on step 2.

This means that some methods might be used in step 2 but not step 1 (if they report all prediction metrics, but no uncertainty metrics, for example). Thus, they influence the rankings (step 2) but disappear (in step 1), leading to inconsistent rankings.

For the same reason, the previous computations for the number of times each model was the best model were inconsistent: for example, we might see a total of 9 "wins" across models, despite there being 10 metrics.

This behavior is fixed by the new CL, which introduces an argument `drop_incomplete_measurements`; if `True`, all rankings are only computed on models that report all metrics. If `False`, all models are kept; when no metrics in a category are reported, the corresponding model's ranking is `NaN`.

PiperOrigin-RevId: 442157194