Fixes a non-determinism bug in the active learning SubsetDatasetBuilder.
Previously, the read_config variable passed in to the SubsetDatasetBuilder was
replaced with an empty ReadConfig object. When shuffle_files=True, the
SubsetdatasetBuilder was shuffling examples using the read_config.shuffle_seed,
which is set to `None` by default. In this case, TensorFlow generates a new
random seed that is non-deterministic, leading to non-deterministic shuffling
during training in the `finetune` function.

This fixes the read_config issue and removes the shuffling from the
SubsetDatasetBuilder since shuffle_files is intended for shuffling raw files.
Since the SubsetDatasetBuilder generates a dataset using a specific set of
example ids, file shuffling is not needed for the underlying builder either.

PiperOrigin-RevId: 442795675