Allow customize `label_key` for the TFDS data loader.
Currently, we hardcode the `label_key` in the TFDS dataset to be 'label'. This can lead to error for datasets whose label key is not 'label' (e.g., 'toxicity' for toxic comments). In particular, calling `dataset.info` will return an error.

This change allows user to set the `label_key` so it works correct for their datasets.

PiperOrigin-RevId: 443107058