Make our AL acquisition functions more amenable to stochastic acquisition without temperature tuning
For BALD and Entropy, coldness = 1 works quite well. Those scores are non-negative and in a range of [0,1.5], from experience.
Update margin and MSP to use `1-scores` instead of `-scores`, which makes these scores non-negative, too.

Rename `power_score_acquisition` to `stochastic_score_acquisition` (as it is power acquisition when using log scores and softmax acquisition without log).
Use log scores. The temperature ablations in the Stochastic Acquisition paper show that Power* is sensitive to temperature changes (performs well).

This is because a score of 0 is interpret as not informative at all. Which margin and MSP adhere to in these proposed changes.